+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## Unreleased
+
+* Add `site` configuration parameter to `.WriteTo.DatadogLogs(...)` and a corresponding
+  `Site` property on `DatadogConfiguration`. The site is used to derive the HTTP intake
+  URL (`https://http-intake.logs.{site}`). Defaults to `datadoghq.com`. Valid values
+  include `datadoghq.com`, `datadoghq.eu`, `us3.datadoghq.com`, `us5.datadoghq.com`,
+  `ap1.datadoghq.com`, `ddog-gov.com`. An explicit `url` continues to take precedence.
+
 ## 0.6.1
 
 * Add `DatadogConfiguration.ResolveHostIfMissing` (and env `DD_LOGS_SINK_RESOLVE_HOST`) to provide a default `host` value when the sink `Host` option is unset. Uses `Environment.MachineName` if available, will fallback to `COMPUTERNAME`/`HOSTNAME` on `netstandard1.x`